refactor(agents-migrator): harden remapAgentPrefixIds error handling
Move remapAgentPrefixIds to its own module (decouples it from AgentsMigrator
and lets future callers depend on just the remap path).

Tighten error handling along the FK-OFF window:
- Add PRAGMA foreign_key_check inside the transaction so any FK update that
  missed rows aborts with detail before the COMMIT, instead of being papered
  over by the eventual PRAGMA foreign_keys = ON.
- Upgrade ROLLBACK / DETACH / PRAGMA-restore failure logs from warn to error,
  with explicit "DB may be in an inconsistent state" wording.
- If PRAGMA foreign_keys = ON itself fails, surface the failure through a
  pendingError box and rethrow after cleanup. Throwing from finally would
  trip no-unsafe-finally and overwrite the original error context, so the
  cleanup runs as straight-line code with the original failure preferred over
  the cleanup failure.

Update tests: AgentsMigrator mocks now expose db.all to satisfy the new
foreign_key_check probe; remapAgentPrefixIds.test.ts imports from its new
module location.
