+++ src/main/data/migration/v2/migrators/__tests__/remapAgentPrefixIds.test.ts
@@ -6,7 +6,7 @@ import { setupTestDatabase } from '@test-helpers/db'
 import { eq, sql } from 'drizzle-orm'
 import { describe, expect, it } from 'vitest'
 
-import { remapAgentPrefixIds } from '../AgentsMigrator'
+import { remapAgentPrefixIds } from '../remapAgentPrefixIds'
 
 const UUID_PATTERN = /^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/
